Start work on binary data
Compiles, but doesn't run yet (because I need to improve my NSCopying skills)

+++ plugins_tools/eid-viewer/conversions/convworker.h
@@ -16,7 +16,7 @@ public:
 	virtual std::string convert(std::string original) { return convert(original.c_str()); };
 	virtual void* convert(const char* original, int* len_return) {
 		std::string cv = convert(original);
-		*len_return = cv.length();
+		*len_return = (int)cv.length();
 		return strdup(cv.c_str());
 	}
 };
